+++ test/unit/fuzz-tools.test.js
@@ -235,6 +235,7 @@ import * as environV242 from '../../lib/environ-v242.js';
 import * as entsleepV243 from '../../lib/entsleep-v243.js';
 import * as sportsmskV244 from '../../lib/sportsmsk-v244.js';
 import * as hemedermV245 from '../../lib/hemederm-v245.js';
+import * as ibdV246 from '../../lib/ibd-v246.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -414,6 +415,7 @@ const MODULES = {
   'entsleep-v243.js': entsleepV243,
   'sportsmsk-v244.js': sportsmskV244,
   'hemederm-v245.js': hemedermV245,
+  'ibd-v246.js': ibdV246,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];